requirements.txt
Django>=3,<4.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
psycopg2>=2.8
pandas<=2.0.3
scikit-learn<=1.3.0

requests~=2.31.0
numpy>=1.24.3

psutil~=5.9.0
pyyaml~=6.0

confluent-kafka>=1.5.0
kafka-schema-registry==0.1.2
jsonschema<4.0


